refactor: use libro-common instead of mana-common
