+++ packages/libro-lab/package.json
@@ -60,6 +60,7 @@
     "@difizen/libro-widget": "^0.3.4",
     "@difizen/libro-l10n": "^0.3.4",
     "@difizen/libro-ai-native": "^0.3.4",
+    "@difizen/libro-common": "^0.3.4",
     "@difizen/mana-app": "latest",
     "@difizen/mana-l10n": "latest",
     "@difizen/mana-common": "latest",
