+++ packages/libro-language-client/src/common/vscodeAdaptor/vscodeAdaptor.ts
@@ -1,5 +1,5 @@
+import { Emitter } from '@difizen/libro-common/mana-common';
 import type { Syringe } from '@difizen/mana-app';
-import { Emitter } from '@difizen/mana-common';
 import * as monaco from '@difizen/monaco-editor-core';
 import type { Disposable } from 'vscode';
 import { URI } from 'vscode-uri';
